Fixes issue #2437 - Split out external/internal classes for WelcomeFileExtension
