+++ extension/welcomefile/pom.xml
@@ -16,6 +16,13 @@
     <build>
         <plugins>
             <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <configuration>
+                    <argLine>@{argLine} --add-opens cloud.piranha.extension.welcomefile/cloud.piranha.extension.welcomefile.internal=ALL-UNNAMED</argLine>
+                </configuration>
+            </plugin>
+            <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
                 <executions>
@@ -45,7 +52,7 @@
                         </configuration>
                     </execution>
                 </executions>            
-            </plugin>            
+            </plugin>
         </plugins>
     </build>
     <dependencies>
